Create BaseButton and focus button on click to address Safari bug
