+++ .changeset/fuzzy-years-hug.md
@@ -1,0 +1,5 @@
+---
+'@spark-web/button': minor
+---
+
+Create BaseButton and focus button on click to address Safari bug
