+++ packages/button/src/Button.tsx
@@ -2,10 +2,9 @@ import { VisuallyHidden } from '@spark-web/a11y';
 import { Box } from '@spark-web/box';
 import type { IconProps } from '@spark-web/icon';
 import { Spinner } from '@spark-web/spinner';
-import { buildDataAttributes } from '@spark-web/utils/internal';
-import type { MouseEvent as ReactMouseEvent } from 'react';
 import { forwardRef } from 'react';
 
+import { BaseButton } from './BaseButton';
 import { resolveButtonChildren } from './resolveButtonChildren';
 import type { CommonButtonProps, NativeButtonProps } from './types';
 import { useButtonStyles } from './useButtonStyles';
@@ -52,17 +51,17 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
       'aria-describedby': ariaDescribedBy,
       'aria-expanded': ariaExpanded,
       data,
-      disabled = false,
+      disabled,
       id,
       loading = false,
       onClick,
       prominence = 'high',
       size = 'medium',
       tone = 'primary',
-      type = 'button',
+      type,
       ...props
     },
-    ref
+    forwardedRef
   ) => {
     const iconOnly = Boolean(props.label);
     const buttonStyleProps = useButtonStyles({
@@ -77,26 +76,20 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
 
     const variant = variants[prominence][tone];
 
-    /**
-     * handle "disabled" behaviour w/o disabling buttons
-     * @see https://axesslab.com/disabled-buttons-suck/
-     */
-    const handleClick = getPreventableClickHandler(onClick, isDisabled);
-
     return (
-      <Box
+      <BaseButton
+        {...buttonStyleProps}
         aria-controls={ariaControls}
         aria-describedby={ariaDescribedBy}
         aria-disabled={isDisabled}
         aria-expanded={ariaExpanded}
         aria-label={props.label}
-        as="button"
+        data={data}
+        disabled={isDisabled}
         id={id}
-        onClick={handleClick}
-        ref={ref}
+        onClick={onClick}
+        ref={forwardedRef}
         type={type}
-        {...buttonStyleProps}
-        {...(data ? buildDataAttributes(data) : undefined)}
       >
         {resolveButtonChildren({
           ...props,
@@ -106,31 +99,11 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
           tone,
         })}
         {isLoading && <Loading tone={variant?.textTone} />}
-      </Box>
+      </BaseButton>
     );
   }
 );
 Button.displayName = 'Button';
-
-/**
- * Prevent click events when the component is "disabled".
- * Note: we don't want to actually disable a button element for several reasons.
- * One being because that would prohibit the use of tooltips.
- */
-export function getPreventableClickHandler(
-  onClick: NativeButtonProps['onClick'],
-  disabled: boolean
-) {
-  return function handleClick(
-    event: ReactMouseEvent<HTMLButtonElement, MouseEvent>
-  ) {
-    if (disabled) {
-      event.preventDefault();
-    } else {
-      onClick?.(event);
-    }
-  };
-}
 
 function Loading({ tone }: { tone?: IconProps['tone'] }) {
   return (
