+++ packages/button/src/index.ts
@@ -1,7 +1,9 @@
+export { BaseButton } from './BaseButton';
 export { Button } from './Button';
 export { ButtonLink } from './ButtonLink';
 
 // types
 
+export type { BaseButtonProps } from './BaseButton';
 export type { ButtonProps } from './Button';
 export type { ButtonLinkProps } from './ButtonLink';
